examples/rag-example/requirements.txt
streamlit==1.37.1
langchain==0.3.23
langchain-openai==0.3.21
python-dotenv==1.1.0
sqlalchemy==2.0.34
pandas==2.2.2
PyPDF2==3.0.1
faiss-cpu==1.10.0